% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annual_abundance}
\alias{annual_abundance}
\title{Annual abundance of outmigrants}
\format{
A named list of four named vectors:
\describe{
  \item{List names}{"Fall", "LateFall", "Spring", "Winter"}
  \item{Vector names}{"1997", "1998", ..., "2011"}
}
}
\usage{
annual_abundance
}
\description{
Annual outmigrant abundance that enters the model in the Sacramento River at Knights Landing
for each juvenile Chinook Salmon run in each year of the 15-yr model period.
}
\examples{
annual_abundance[["Fall"]][["2000"]]

}
\keyword{datasets}
