% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sar.R
\name{calculate_sar}
\alias{calculate_sar}
\title{Calculate SAR}
\usage{
calculate_sar(sim_name)
}
\arguments{
\item{sim_name}{Name of simulation, e.g., 'Baseline' would be used to match files named as Baseline-Reps-1-50.rds, Baseline-Reps-51-100.rds, etc.}
}
\description{
Calculate SAR as the juvenile abundance at Knights Landing divided by adult ocean returners
}
