% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_sar_difference.R
\name{calculate_sar_difference}
\alias{calculate_sar_difference}
\title{Calculate SAR difference}
\usage{
calculate_sar_difference(sim_name)
}
\arguments{
\item{sim_name}{Name of simulation, e.g., 'Baseline' would be used to match files named as Baseline-Reps-1-50.rds, Baseline-Reps-51-100.rds, etc.}
}
\description{
Calculate difference in SAR between Yolo Bypass and Sacramento River routes where SAR is defined
as the juvenile abundance at Fremont Weir divided by adult ocean returners
}
