% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_ocean_year_type.R
\name{draw_ocean_year_type}
\alias{draw_ocean_year_type}
\title{Draw ocean year type}
\usage{
draw_ocean_year_type()
}
\description{
Determine whether ocean year type is length or intercept; used in ocean survival relationship
}
