% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{extract_results}
\alias{extract_results}
\title{Extract selected columns from raw model output}
\usage{
extract_results(sim_name, sac_columns, yolo_columns)
}
\arguments{
\item{sim_name}{Name of simulation, e.g., 'Baseline' would be used to match files named as Baseline-Reps-1-50.rds, Baseline-Reps-51-100.rds, etc.}

\item{sac_columns}{Vector of Sac column names to extract}

\item{column}{Vector of Yolo column names to extract}
}
\description{
Returns a list with two dataframes (Sac and Yolo).
}
