% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flood_duration}
\alias{flood_duration}
\title{Flood duration}
\format{
A list of 15 vectors of flood durations. One list element for each water year (1997-2011).
}
\usage{
flood_duration
}
\description{
Number of days where Yolo Bypass flows exceeded 4,000 cfs in a window of 120 days
from the date that cohort entered the Yolo Bypass.
}
\examples{
flood_duration[["1998"]][105:115]

}
\keyword{datasets}
