% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodplain_growth.R
\name{floodplain_growth}
\alias{floodplain_growth}
\title{Floodplain growth}
\usage{
floodplain_growth(
  wet_weight,
  temp,
  duration,
  params = floodplain_growth_parameters
)
}
\arguments{
\item{wet_weight}{Wet weight (g) at start of rearing period}

\item{temp}{Mean temperature (ºC) during the rearing period}

\item{duration}{Duration (days) of rearing period}

\item{params}{Parameters for the Ratkowsky growth model}
}
\description{
Returns final wet_weight given initial wet_weight and mean temperture during growth period; uses parameters from Perry et al. 2015
}
\examples{
floodplain_growth(10, 2, 20)
floodplain_growth(10, 20, 20)

}
