% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{floodplain_temperature}
\alias{floodplain_temperature}
\title{Floodplain temperature}
\format{
A nested named list of 15 vectors of daily temperature values. One list element for each location (Delta, Yolo) and one sub-list element for each water year (1997-2011).
}
\usage{
floodplain_temperature
}
\description{
Daily temperature (ºC) on the floodplain (where floodplain is inclusive of any off-channel rearing habitat)
}
\examples{
floodplain_temperature[["Delta"]][["2000"]][100:110]

}
\keyword{datasets}
