% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{floodplain_temperature_difference}
\alias{floodplain_temperature_difference}
\title{Floodplain temperature difference}
\format{
A list of 2 vectors of 366 elements:
\describe{
  \item{DOY}{day of year}
  \item{Value}{difference in mean daily temperature (ºC)}
}
}
\usage{
floodplain_temperature_difference
}
\description{
A dataset containing the difference between in temperature between the Toe Drain and floodplain
}
\keyword{datasets}
