% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freeport_flow}
\alias{freeport_flow}
\title{Freeport Flow}
\format{
A list of 15 vectors of flow values. One list element for each water year (1997-2011).
}
\usage{
freeport_flow
}
\description{
Flow (cfs) in the Sacramento River at Freeport.
}
\examples{
freeport_flow[["1998"]][105:115]

}
\keyword{datasets}
