% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freeport_temperature}
\alias{freeport_temperature}
\title{Freeport temperature}
\format{
A list of 15 vectors of daily temperature values. One list element for each water year (1997-2011).
}
\usage{
freeport_temperature
}
\description{
Daily temperature (ºC) in the Sacramento River at Freeport
}
\examples{
freeport_temperature[["2000"]][100:110]

}
\keyword{datasets}
