% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fremont_weir_proportion}
\alias{fremont_weir_proportion}
\title{Proportion of fish entrained onto the Yolo Bypass at Fremont Weir}
\format{
A list of 15 vectors of daily entrainment proportions. One list element for each water year (1997-2011).
}
\usage{
fremont_weir_proportion
}
\description{
Proportion of fish entering the Yolo Bypass for each day of the 15-yr model period.
}
\examples{
fremont_weir_proportion[["1998"]][105:115]

}
\keyword{datasets}
