% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_floodplain_temperature.R
\name{get_floodplain_temperature}
\alias{get_floodplain_temperature}
\title{Get floodplain temperature}
\usage{
get_floodplain_temperature()
}
\description{
Returns floodplain temperature (sum of Toe Drain/Freeport temperature and floodplain temperature difference);
used to adjust floodplain_temperature_difference and recalculate floodplain temperature
}
