% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_water_year.R
\name{get_water_year}
\alias{get_water_year}
\title{Water year}
\usage{
get_water_year(model_day)
}
\arguments{
\item{model_day}{Model day where 1 is 1996-10-02, 365 is 1997-10-01, 730 is 1998-10-01, etc.}
}
\description{
Get water year from model day.
}
\examples{
get_water_year(c(1, 365, 730))

}
