% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wy_model_days.R
\name{get_wy_model_days}
\alias{get_wy_model_days}
\title{Day of year}
\usage{
get_wy_model_days(water_year)
}
\arguments{
\item{water_year}{Water year from the simulation period, 1997-2011}
}
\description{
Get the model days in a water year
}
\examples{
get_wy_model_days(1997)
get_wy_model_days(1998)
get_wy_model_days(2004)

}
