% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wy_yday.R
\name{get_wy_yday}
\alias{get_wy_yday}
\title{Day of water year}
\usage{
get_wy_yday(model_day)
}
\arguments{
\item{model_day}{Model day where 1 is 1996-10-02, 365 is 1997-10-01, 730 is 1998-10-01, etc.}
}
\description{
Get day of water year from model day.
}
\examples{
get_wy_yday(c(1, 365, 730))

}
