% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{growth}
\alias{growth}
\title{Growth}
\usage{
growth(wet_weight, temp, duration)
}
\arguments{
\item{wet_weight}{Wet weight (g) at start of rearing period}

\item{temp}{Mean temperature (ºC) during the rearing period}

\item{duration}{Duration (days) of rearing period}
}
\description{
Returns final wet_weight given initial wet_weight and mean temperture during growth period; uses parameters from Perry et al. 2015
}
\examples{
growth(10, 2, 20)
growth(10, 20, 20)

}
