% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_parameters}
\alias{growth_parameters}
\title{Growth parameters}
\format{
A named vector:
\describe{
  \item{b}{Allometric growth exponent}
  \item{d,g}{Shape parameters for the Ratkowsky growth model}
  \item{TL,TU}{Lower and upper temperature (ºC) limits for growth}
}
}
\usage{
growth_parameters
}
\description{
Parameters for Perry et al. 2015 growth model
}
\keyword{datasets}
