% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_cohort.R
\name{initialize_cohort}
\alias{initialize_cohort}
\title{Initialize cohort}
\usage{
initialize_cohort(
  run = c("Fall", "LateFall", "Winter", "Spring"),
  model_day,
  abundance,
  fork_length
)
}
\arguments{
\item{run}{Run of cohort}

\item{model_day}{Day that cohort enters the model}

\item{abundance}{Abundance at Knights Landing}

\item{fork_length}{Fork length (mm) at Knights Landing}
}
\description{
Initialize data structure for tracking cohort through model
}
