% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inundated_sqkm}
\alias{inundated_sqkm}
\title{Inundated area}
\format{
A list of 15 vectors of flow values. One list element for each water year (1997-2011).
}
\usage{
inundated_sqkm
}
\description{
Area (sq. km) inundated (depth > 0) on the Yolo Bypass for each day of the 15-yr model period.
}
\examples{
inundated_sqkm[["1998"]][105:115]

}
\keyword{datasets}
