% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap_year.R
\name{is_leap_year}
\alias{is_leap_year}
\title{Leap year}
\usage{
is_leap_year(year)
}
\arguments{
\item{year}{Calendar year}
}
\description{
Check if a year is a leap year
}
\examples{
is_leap_year(2000:2004)

}
