% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knights_landing_fl_params}
\alias{knights_landing_fl_params}
\title{Knights Landing fork length parameters}
\format{
Nested, named lists of vectors of daily values. One list element for each Chinook run (Fall, LateFall, Spring, Winter),
one sub-list element for each water year (1997-2011), and one sub-sub-list for each parameter (MeanLog, SDLog):
}
\usage{
knights_landing_fl_params
}
\description{
Parameters used to draw fork lengths from lognormal distribution for cohorts entering the model
in the Sacramento River at Knights Landing on each day of the 15-yr model period.
}
\examples{
knights_landing_fl_params[["Spring"]][["1998"]][["SDLog"]][105:115]

}
\keyword{datasets}
