% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knights_landing_timing}
\alias{knights_landing_timing}
\title{Knights Landing entry timing}
\format{
Nested, named lists of vectors of daily values. One list element for each Chinook run (Fall, LateFall, Spring, Winter) and one sub-list element for each water year (1997-2011):
}
\usage{
knights_landing_timing
}
\description{
Proportion of the population that enters the model in the Sacramento River at Knights Landing on each day of the 15-yr model period.
}
\examples{
knights_landing_timing[["Spring"]][["1998"]][105:115]

}
\keyword{datasets}
