% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_weight.R
\name{length_weight}
\alias{length_weight}
\title{Length-weight relationshiop}
\usage{
length_weight(fork_length, params = length_weight_parameters)
}
\arguments{
\item{fork_length}{Fork length (mm)}

\item{params}{Parameters for length-weight relationship}
}
\description{
Return wet weight (g) for a given fork length (mm)
}
\examples{
length_weight(30:120)

}
