% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{length_weight_parameters}
\alias{length_weight_parameters}
\title{Length-weight parameters}
\format{
A named vector:
\describe{
  \item{a}{Constant}
  \item{b}{Scaling exponent}
}
}
\usage{
length_weight_parameters
}
\description{
Parameters for an allometric relationship from Tiffan et al. (2014) for converting
between weight (g) and length (mm).
}
\keyword{datasets}
