% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic.R
\name{logistic}
\alias{logistic}
\title{logistic function}
\usage{
logistic(x, a, b, c, d)
}
\arguments{
\item{x}{Independent variable}

\item{a}{Maximum value}

\item{b}{Steepness of curve}

\item{c}{Inflection point}

\item{d}{Minimum value}
}
\description{
Four parameter logistic function
}
