% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ocean_survival_parameters}
\alias{ocean_survival_parameters}
\title{Ocean survival parameters}
\format{
A named list ("intercept" or "length") of named vectors:
\describe{
  \item{inter}{Intercept}
  \item{slope}{Slope}
  \item{phi}{Dispersion}
}
}
\usage{
ocean_survival_parameters
}
\description{
Parameters for two beta-binomial regression models. The "length" model includes fork length as a predictor.
The "intercept" model is an intercept-only model with slope equal to zero.
}
\keyword{datasets}
