% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_year_type.R
\name{ocean_year_type}
\alias{ocean_year_type}
\title{Ocean year type}
\usage{
ocean_year_type()
}
\description{
Determine whether ocean year type is length or intercept; used in ocean survival relationship
}
