% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage.R
\name{passage}
\alias{passage}
\title{Passage survival and time from Fremont Weir to Chipps Island}
\usage{
passage(
  water_year_string,
  date_index,
  abundance,
  fork_length,
  route = c("Sac", "Yolo")
)
}
\arguments{
\item{water_year_string}{Water year (1997-2011) as a string}

\item{date_index}{Index of date in a water year that cohort begins passage; in all years except WY1997, equivalent to day of water year}

\item{abundance}{Abundance of cohort on day route entered}

\item{fork_length}{Fork length (mm) at Fremont Weir}

\item{route}{Route: Sacramento River (Sac) or Yolo Bypass (Yolo)}
}
\description{
Passage survival and time (days) from Fremont Weir to Chipps Island based on fork length, flow, and route
}
