% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_growth.R
\name{passage_growth}
\alias{passage_growth}
\title{Passage growth}
\usage{
passage_growth(wet_weight, model_day, duration)
}
\arguments{
\item{wet_weight}{Wet weight (g) at start of rearing period}

\item{model_day}{Model day when rearing was initiated}

\item{duration}{Duration (days) of rearing period}
}
\description{
Returns final wet_weight given initial wet_weight, model day, and duration
}
