% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_time.R
\name{passage_time}
\alias{passage_time}
\title{Passage time from Fremont Weir to Chipps}
\usage{
passage_time(fork_length, flow, route)
}
\arguments{
\item{fork_length}{Fork length (mm) at Fremont Weir}

\item{flow}{Flow (cfs) at Freeport on day route entered}

\item{route}{Route: Sacramento River (Sac) or Yolo Bypass (Yolo)}
}
\description{
Passage time (days) from Fremont Weir to Chipps Island based on fork length, flow, and route
}
