% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_linear_cdf.R
\name{piecewise_linear_cdf}
\alias{piecewise_linear_cdf}
\title{Piecewise-linear cumulative distribution function}
\usage{
piecewise_linear_cdf(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Transform numeric vector, x, to create a piecewise-linear CDF
with a mean and variance equal to the unbiased sample mean and variance of x.
}
\examples{
get_wy_yday(c(1, 365, 730))

}
