% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_probability.R
\name{rearing_probability}
\alias{rearing_probability}
\title{Rearing probability}
\usage{
rearing_probability(fork_length, params = rearing_probability_parameters)
}
\arguments{
\item{fork_length}{Fork length (mm) at Fremont Weir}

\item{params}{Parameters for four-parameter logistic function}
}
\description{
Rearing probability on the Yolo Bypass based on hypothetical logistic function
}
\examples{
rearing_probability(70)
rearing_probability(150)

}
