% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rearing_proportion_parameters}
\alias{rearing_proportion_parameters}
\title{Rearing proportion parameters}
\format{
A named vector:
\describe{
  \item{min}{Minimum rearing proportion parameter in logistic function}
  \item{max}{Maximum rearing proportion parameter in logistic function}
  \item{inflection}{Inflection parameter in logistic function}
  \item{steepness}{Steepness parameter in logistic function}
}
}
\usage{
rearing_proportion_parameters
}
\description{
Parameters for a logistic function to determine the proportion of an entrained cohort that rears on the Yolo Bypass as a function of fork length.
}
\keyword{datasets}
