% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_status.R
\name{rearing_status}
\alias{rearing_status}
\title{Rearing status}
\usage{
rearing_status(fork_length)
}
\arguments{
\item{fork_length}{Fork length (mm) at Fremont Weir}
}
\description{
Rearing status for cohorts on the Yolo Bypass (0 = not rearing; 1 = rearing); note: no deterministic option
}
\examples{
rearing_status(70)
rearing_status(150)

}
