% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_survival.R
\name{rearing_survival}
\alias{rearing_survival}
\title{Rearing survival}
\usage{
rearing_survival(
  water_year_string,
  date_index,
  abundance,
  duration,
  location = c("Delta", "Yolo"),
  sim_type
)
}
\arguments{
\item{water_year_string}{Water year (1997-2011) as a string}

\item{date_index}{Index of date in a water year at start of rearing period; in all years except WY1997, equivalent to day of water year}

\item{abundance}{Abundance at start of rearing period}

\item{duration}{Number of days spent rearing}

\item{location}{Rearing location: Delta or Yolo}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Rearing survival based on rearing time and daily survival
}
