% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_time.R
\name{rearing_time}
\alias{rearing_time}
\title{Floodplain rearing time}
\usage{
rearing_time(model_day, scenario, sim_type)
}
\arguments{
\item{model_day}{Model day when rearing was initiated}

\item{scenario}{Scenario: Exg, Alt01, Alt04b, Alt04, Alt05, Alt06}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Floodplain rearing time after adjusting indirectly for passage time and directly for temperature
}
