% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_time_delta.R
\name{rearing_time_delta}
\alias{rearing_time_delta}
\title{Delta rearing time}
\usage{
rearing_time_delta(water_year_string, date_index, passage_time, fork_length)
}
\arguments{
\item{water_year_string}{Water year (1997-2011) as a string}

\item{date_index}{Index of date in a water year that cohort begins Delta rearing; in all years except WY1997, equivalent to day of water year}

\item{passage_time}{Passge time}

\item{fork_length}{Average fork length (mm) of cohort}
}
\description{
Delta rearing time after adjusting for passage time and temperature
}
