% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rearing_time_parameters}
\alias{rearing_time_parameters}
\title{Rearing time parameters}
\format{
A named list ("Delta" or "Yolo") of named vectors:
\describe{
  \item{inter}{Intercept}
  \item{slope}{Slope for flood duration predictor (Yolo)}
  \item{flow}{Coefficient for flow predictor (Delta)}
  \item{fork_length}{Coefficient for fork length predictor (Delta)}
  \item{theta}{Dispersion}
  \item{passage_days}{Number of passage days to subtract from overall residence time (Yolo)}
  \item{theta}{Temperature threshold (ºC) at which fish stop rearing.}
}
}
\usage{
rearing_time_parameters
}
\description{
Parameters for negative-binomial regression models for the Delta and Yolo Bypass.
}
\keyword{datasets}
