% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_time_proportion.R
\name{rearing_time_proportion}
\alias{rearing_time_proportion}
\title{Rearing time proportion}
\usage{
rearing_time_proportion(fork_length)
}
\arguments{
\item{fork_length}{Fork length (mm) at Fremont Weir}
}
\description{
Proportion of drawn rearing time that cohort \emph{potentially} rears on the Yolo Bypass based on hypothetical logistic function;
actual rearing time might be truncated by temperature thresholds
}
\examples{
rearing_time_proportion(70)
rearing_time_proportion(150)

}
