% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_time_yolo.R
\name{rearing_time_yolo}
\alias{rearing_time_yolo}
\title{Yolo Bypass rearing time}
\usage{
rearing_time_yolo(water_year_string, date_index)
}
\arguments{
\item{water_year_string}{Water year (1997-2011) as a string}

\item{date_index}{Index of date in a water year that cohort begins Yolo rearing; in all years except WY1997, equivalent to day of water year}
}
\description{
Yolo Bypass rearing time after adjusting indirectly for passage time and directly for temperature
}
