% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_model_inputs.R
\name{reset_model_inputs}
\alias{reset_model_inputs}
\title{Reset model inputs}
\usage{
reset_model_inputs()
}
\description{
Reset model input data and parameters by removing any items with the same names from the global environment.
}
