% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_one_rep.R
\name{run_one_rep}
\alias{run_one_rep}
\title{Run one replicate of the model}
\usage{
run_one_rep(water_year_string, chinook_run)
}
\arguments{
\item{water_year_string}{Water year (1997-2011) as a string}

\item{chinook_run}{Run timing classification: Fall, LateFall, Winter, Spring}
}
\description{
Run one replicate of the model; a replicate is a water_year/chinook_run combination
}
