% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run simulation}
\usage{
run_simulation(cores = parallel::detectCores())
}
\description{
Run simulation based on simulation parameters. They typical workflow for running simulation is to reset_model_inputs(),
set (in the global environment) model input data and parameter for a simulation, and then use run_simulation().
}
