% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_parameters}
\alias{simulation_parameters}
\title{Simulation parameters}
\format{
A named list containing the following parameters:
\describe{
  \item{name}{Simulation name. Used for naming output files.}
  \item{sim_type}{Simulation type: "stochastic" or "deterministic".}
  \item{reps}{Vector of replicate numbers, e.g., 1:50, 51:100, etc.}
  \item{water_years}{Numeric vector of water years (1997-2011) to include in a simulation. Default is 1997:2011.}
  \item{chinook_runs}{Chinook runs to include in a simulation. Default is c("Fall", "LateFall", "Spring", "Winter").}
  \item{ocean_year_type_probability}{Probability of ocean survival being based on fork length relationship.}
}
}
\usage{
simulation_parameters
}
\description{
Parameters used in `run_simulation()`.
}
\keyword{datasets}
