% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_quantiles.R
\name{summarise_quantiles}
\alias{summarise_quantiles}
\title{Summarise quantiles of replicates in a grouped dataframe}
\usage{
summarise_quantiles(grouped_df, column)
}
\arguments{
\item{grouped_df}{Grouped data frame (usually formed by dplyr::group_by)}

\item{column}{Column to summarise}
}
\description{
Common operation to summarise model results across replicates.
}
