% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_adjustment.R
\name{temperature_adjustment}
\alias{temperature_adjustment}
\title{Adjust rearing time based on temperature}
\usage{
temperature_adjustment(water_year_string, date_index, duration, location)
}
\arguments{
\item{water_year_string}{Water year (1997-2011) as a string}

\item{date_index}{Index of date in a water year at start of rearing period; in all years except WY1997, equivalent to day of water year}

\item{duration}{Maximum numbr of rearing days}

\item{location}{Rearing location: Yolo or Delta}
}
\description{
Rearing time after potentially truncating based on temperature threshold
}
