% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toe_drain_temperature}
\alias{toe_drain_temperature}
\title{Toe Drain temperature}
\format{
A list of 15 vectors of daily temperature values. One list element for each water year (1997-2011).
}
\usage{
toe_drain_temperature
}
\description{
Daily temperature (ºC) in the Toe Drain at the screw trap
}
\examples{
toe_drain_temperature[["2000"]][100:110]

}
\keyword{datasets}
