% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_time.R
\name{travel_time}
\alias{travel_time}
\title{Draw travel time from Fremont Weir to Chipps}
\usage{
travel_time(
  fork_length,
  flow,
  route = c("Sac", "Yolo"),
  sim_type = c("deterministic", "stochastic"),
  params = telemetry_parameters
)
}
\arguments{
\item{fork_length}{Fork length (mm) at estuary entry (Chipps Island)}

\item{flow}{Flow (cfs) at Freeport on day route entered}

\item{sim_type}{Simulation type: deterministic or stochastic}

\item{params}{Parameters from Bayesian model of survival and travel time based on acoustic telemetry data}
}
\description{
Draw a random travel time (days) from Fremont Weir to Chipps Island based on fork length, flow, and route
}
\examples{
travel_time(70, 60000, "Sac")
travel_time(70, 60000, "Yolo")

}
