% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_year_type}
\alias{water_year_type}
\title{Water year type}
\format{
A data frame with 2 columns and 15 rows:
\describe{
  \item{WaterYear}{Water year: 1997-2011}
  \item{WaterYearType}{Water year classification. Above Normal and Below Normal include newline characters for nicer display when plotting.}
}
}
\usage{
water_year_type
}
\description{
Water year classifications for the Sacramento River from http://cdec.water.ca.gov/reportapp/javareports?name=WSIHIST.
}
\keyword{datasets}
