% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_length.R
\name{weight_length}
\alias{weight_length}
\title{Length-weight relationshiop}
\usage{
weight_length(wet_weight, params = length_weight_parameters)
}
\arguments{
\item{params}{Parameters for length-weight relationship}

\item{fork_length}{Fork length (mm)}
}
\description{
Return wet weight (g) for a given fork length (mm)
}
\examples{
length_weight(30:120)

}
