% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_abundance.R
\docType{data}
\name{annual_abundance}
\alias{annual_abundance}
\title{Annual abundance of outmigrants}
\format{A data frame with 15 rows and 5 columns:
\describe{
  \item{WaterYear}{defined as Oct 1st to Sept 30th}
  \item{Winter}{winter-run annual abundance}
  \item{Spring}{spring-run annual abundance}
  \item{Fall}{fall-run annual abundance}
  \item{LateFall}{late-fall-run annual abundance}
}}
\usage{
annual_abundance
}
\description{
A dataset containing the annual outmigrant abundance that enters the model
in the Sacramento River at Knights Landing for each juvenile Chinook Salmon run
in each year of the 15-yr model period.
}
\keyword{datasets}
