% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrainment.R
\name{entrainment}
\alias{entrainment}
\title{Number of fish entrained at Fremont Weir}
\usage{
entrainment(
  scenario = c("Exg", "Alt01", "Alt04b", "Alt04", "Alt05", "Alt06"),
  model_day,
  abundance
)
}
\arguments{
\item{scenario}{Scenario: Exg, Alt01, Alt04b, Alt04, Alt05, Alt06}

\item{model_day}{Day that cohort reaches Fremont Weir}

\item{abundance}{Abundance of cohort at Fremont Weir}
}
\description{
Draw random number of fish entrained at Fremont Weir based on scenario, model day, and abundance at Fremont Weir
}
