% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrainment.R
\name{entrainment}
\alias{entrainment}
\title{Fish entrainment at Fremont Weir}
\usage{
entrainment(model_day, abundance, sim_type)
}
\arguments{
\item{model_day}{Day that cohort reaches Fremont Weir}

\item{abundance}{Abundance of cohort at Fremont Weir}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Draw random number of fish entrained at Fremont Weir based on model day and abundance at Fremont Weir
}
