% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flood_duration}
\alias{flood_duration}
\title{Flood duration}
\format{
A list comprised of 2 vectors of 5477 elements:
\describe{
  \item{Date}{calendar date}
  \item{Value}{flood duration (days)}
}
}
\usage{
flood_duration
}
\description{
Number of days where Yolo Bypass flows exceeded 4,000 cfs in a window of 120 days
from the date that cohort entered the Yolo Bypass.
}
\keyword{datasets}
