% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodplain_growth.R
\name{floodplain_growth}
\alias{floodplain_growth}
\title{Floodplain growth}
\usage{
floodplain_growth(mass, temp, duration, params = floodplain_growth_parameters)
}
\arguments{
\item{mass}{Mass (g) at start of rearing period}

\item{temp}{Mean temperature (ºC) during the rearing period}

\item{duration}{Duration (days) of rearing period}

\item{params}{Parameters for the Ratkowsky growth model}
}
\description{
Returns final mass given initial mass and mean temperture during growth period; uses parameters from Perry et al. 2015
}
\examples{
floodplain_growth(10, 2, 20)
floodplain_growth(10, 20, 20)

}
