% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{floodplain_temperature}
\alias{floodplain_temperature}
\title{Floodplain temperature}
\format{
A list of 2 vectors of 5477 elements:
\describe{
  \item{Date}{calendar date}
  \item{Value}{mean daily temperature (ºC)}
}
}
\usage{
floodplain_temperature
}
\description{
A dataset containing the predicted temperature on the Yolo Bypass floodplain
}
\keyword{datasets}
