% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freeport_flow}
\alias{freeport_flow}
\title{Freeport Flow}
\format{
A list comprised of 2 vectors of 5477 elements:
\describe{
  \item{Date}{calendar date}
  \item{Value}{flow (cfs) at Freeport}
}
}
\usage{
freeport_flow
}
\description{
A dataset containing the modeled flow at Freeport.
}
\keyword{datasets}
