% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{freeport_temperature}
\alias{freeport_temperature}
\title{Freeport temperature}
\format{
A list of 2 vectors of 5477 elements:
\describe{
  \item{Date}{calendar date}
  \item{Value}{mean daily temperature (ºC)}
}
}
\usage{
freeport_temperature
}
\description{
A dataset containing the daily temperature in the Sacramento River at Freeport
}
\keyword{datasets}
