% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fremont_weir_proportion}
\alias{fremont_weir_proportion}
\title{Proportion of fish entrained onto the Yolo Bypass at Fremont Weir}
\format{
A list comprised of 2 vectors of 5477 elements:
\describe{
  \item{Date}{calendar date}
  \item{Value}{proportion of fish entering the Yolo Bypass}
}
}
\usage{
fremont_weir_proportion
}
\description{
A dataset containing the predicted proportion of fish entering the Yolo Bypass
for each day of the 15-yr model period.
}
\keyword{datasets}
