% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_cohort_abundance.R
\name{initial_cohort_abundance}
\alias{initial_cohort_abundance}
\title{Initial cohort abundance}
\usage{
initial_cohort_abundance(water_year, chinook_run, sim_type)
}
\arguments{
\item{water_year}{Water year 1997-2011}

\item{chinook_run}{Run timing classification: Fall, LateFall, Winter, Spring}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Initial cohort abundances based on water year and Chinook run
}
