% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_cohort_abundance.R
\name{initial_cohort_abundance}
\alias{initial_cohort_abundance}
\title{Draw initial cohort abundance}
\usage{
initial_cohort_abundance(
  run = c("Fall", "LateFall", "Winter", "Spring"),
  water_year
)
}
\arguments{
\item{run}{Run timing classification: Fall, LateFall, Winter, Spring}

\item{water_year}{Water year 1997-2011}
}
\description{
Draw a random vector of initial cohort abundances based on run timing, water year, and annual abundance
}
\examples{

}
