% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_cohort_fork_length.R
\name{initial_cohort_fork_length}
\alias{initial_cohort_fork_length}
\title{Draw initial cohort fork length}
\usage{
initial_cohort_fork_length(
  run = c("Fall", "LateFall", "Winter", "Spring"),
  model_day
)
}
\arguments{
\item{run}{Run timing classification: Fall, LateFall, Winter, Spring}

\item{model_day}{Day that cohort enters the model}
}
\description{
Draw initial cohort for fork length based on run timing and model day
}
