% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_cohort_fork_length.R
\name{initial_cohort_fork_length}
\alias{initial_cohort_fork_length}
\title{Initial cohort fork length}
\usage{
initial_cohort_fork_length(model_day, chinook_run, sim_type)
}
\arguments{
\item{model_day}{Day that cohort enters the model}

\item{chinook_run}{Run timing classification: Fall, LateFall, Winter, Spring}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Initial cohort fork length based on model day and Chinook run
}
