% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{knights_landing_fl_params}
\alias{knights_landing_fl_params}
\title{Knights Landing fork length parameters}
\format{A list comprised of 4 lists (one for each run: Fall, LateFall, Spring, Winter) that contain 3 vectors (Date, MeanLog, SDLog) of 5382 elements:
\describe{
  \item{Date}{calendar date}
  \item{MeanLog}{meanlog parameter of lognormal distribution}
  \item{SDLog}{sdlog parameter of lognormal distribution}
}}
\usage{
knights_landing_fl_params
}
\description{
A dataset containing the parameters used to draw random fork lengths for cohorts entering the model
in the Sacramento River at Knights Landing on each day of the 15-yr model period.
}
\keyword{datasets}
