% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_survival.R
\name{ocean_survival}
\alias{ocean_survival}
\title{Number of adults returning from ocean}
\usage{
ocean_survival(fork_length, sim_type = c("deterministic", "stochastic"))
}
\arguments{
\item{fork_length}{Fork length (mm) at estuary entry (Chipps Island)}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Number of adults returning from ocean after three years based on fork length
}
\examples{
ocean_survival(100)

}
