% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_survival.R
\name{ocean_survival}
\alias{ocean_survival}
\title{Draw number of adults returning from ocean}
\usage{
ocean_survival(fork_length)
}
\arguments{
\item{fork_length}{Fork length (mm) at estuary entry (Chipps Island)}
}
\description{
Draw a random number of adults returning from ocean after 3 years based on fork length
}
\examples{
ocean_survival(100)

}
