% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_survival.R
\name{ocean_survival}
\alias{ocean_survival}
\title{Number of adults returning from ocean}
\usage{
ocean_survival(
  mass,
  abundance,
  sim_type = c("deterministic", "stochastic"),
  params = ocean_survival_parameters
)
}
\arguments{
\item{mass}{Average mass (g) of cohort at estuary entry (Chipps Island)}

\item{abundance}{Abundance of cohort at esturary entry}

\item{sim_type}{Simulation type: deterministic or stochastic}

\item{params}{Parameters from betabinomial regression of ocean recoveries versus release weight}
}
\description{
Number of adults returning from ocean after three years based on mass
}
\examples{
ocean_survival(40, 1000)

}
