% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocean_survival.R
\name{ocean_survival}
\alias{ocean_survival}
\title{Number of adults returning from ocean}
\usage{
ocean_survival(wet_weight, abundance, sim_type)
}
\arguments{
\item{wet_weight}{Average wet weight (g) of cohort at estuary entry (Chipps Island)}

\item{abundance}{Abundance of cohort at esturary entry}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Number of adults returning from ocean after three years based on weight
}
\examples{
ocean_survival(40, 1000)

}
