% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage.R
\name{passage}
\alias{passage}
\title{Passage survival and time from Fremont Weir to Chipps Island}
\usage{
passage(
  model_day,
  abundance,
  fork_length,
  route = c("Sac", "Yolo"),
  scenario,
  sim_type
)
}
\arguments{
\item{model_day}{Model day at start of passage}

\item{abundance}{Abundance of cohort on day route entered}

\item{fork_length}{Fork length (mm) at Fremont Weir}

\item{route}{Route: Sacramento River (Sac) or Yolo Bypass (Yolo)}

\item{scenario}{Scenario: Exg, Alt01, Alt04b, Alt04, Alt05, Alt06}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Passage survival and time (days) from Fremont Weir to Chipps Island based on fork length, flow, and route
}
