% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passage_survival.R
\name{passage_survival}
\alias{passage_survival}
\title{Passage survival from Fremont Weir to Chipps Island}
\usage{
passage_survival(
  fork_length,
  flow,
  abundance,
  route = c("Sac", "Yolo"),
  sim_type = c("deterministic", "stochastic"),
  params = telemetry_parameters
)
}
\arguments{
\item{fork_length}{Fork length (mm) at Fremont Weir}

\item{flow}{Flow (cfs) at Freeport on day route entered}

\item{abundance}{Abundance of cohort on day route entered}

\item{sim_type}{Simulation type: deterministic or stochastic}

\item{params}{Parameters from Bayesian model of survival and travel time based on acoustic telemetry data}
}
\description{
Passage survival from Fremont Weir to Chipps Island based on fork length, flow, and route
}
\examples{
passage_survival(70, 60000, 100000, "Sac")
passage_survival(70, 60000, 100000, "Yolo")

}
