% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_abundance.R
\name{rearing_abundance}
\alias{rearing_abundance}
\title{Rearing abundance}
\usage{
rearing_abundance(abundance, fork_length, sim_type)
}
\arguments{
\item{abundance}{Abundance of cohort at Fremot Weir}

\item{fork_length}{Fork length (mm) at Fremont Weir}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Number of individuals in cohort entrained onto Yolo Bypass that actually rears on the Yolo Bypass based on hypothetical logistic function with fork length
}
