% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_survival.R
\name{rearing_survival}
\alias{rearing_survival}
\title{Draw rearing survival}
\usage{
rearing_survival(rearing_time, min, max)
}
\arguments{
\item{rearing_time}{Number of days spent rearing}

\item{min, max}{Lower and upper limits of uniform distribution of daily survival values}
}
\description{
Draw a random rearing survival based on rearing time and daily survival
}
\examples{
rearing_survival(10, 0.95, 0.999)

}
