% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_survival.R
\name{rearing_survival}
\alias{rearing_survival}
\title{Rearing survival}
\usage{
rearing_survival(
  rearing_time,
  sim_type = c("deterministic", "stochastic"),
  params = rearing_survival_parameters
)
}
\arguments{
\item{rearing_time}{Number of days spent rearing}

\item{sim_type}{Simulation type: deterministic or stochastic}

\item{params}{Rearing survival parameters}
}
\description{
Rearing survival based on rearing time and daily survival
}
\examples{
rearing_survival(10)

}
