% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_survival.R
\name{rearing_survival}
\alias{rearing_survival}
\title{Rearing survival}
\usage{
rearing_survival(
  model_day,
  abundance,
  duration,
  location = c("Delta", "Yolo"),
  sim_type
)
}
\arguments{
\item{model_day}{Model day at start of rearing period}

\item{abundance}{Abundance at start of rearing period}

\item{duration}{Number of days spent rearing}

\item{location}{Rearing location: Yolo or Delta}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Rearing survival based on rearing time and daily survival
}
