% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_time_adj.R
\name{rearing_time_adj}
\alias{rearing_time_adj}
\title{Adjusted floodplain rearing time}
\usage{
rearing_time_adj(rearing_time_max, model_day)
}
\arguments{
\item{rearing_time_max}{Maximum rearing duration (days)}

\item{model_day}{Model day when rearing was initiated}
}
\description{
Floodplain rearing time adjusted for fish size and temperature
}
