% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_time_delta.R
\name{rearing_time_delta}
\alias{rearing_time_delta}
\title{Delta rearing time}
\usage{
rearing_time_delta(model_day, passage_time, fork_length, scenario, sim_type)
}
\arguments{
\item{model_day}{Model day when rearing was initiated}

\item{passage_time}{Passge time}

\item{fork_length}{Average fork length (mm) of cohort}

\item{scenario}{Scenario: Exg, Alt01, Alt04b, Alt04, Alt05, Alt06}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Delta rearing time after adjusting for passage time and temperature
}
