% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearing_time_yolo.R
\name{rearing_time_yolo}
\alias{rearing_time_yolo}
\title{Yolo Bypass rearing time}
\usage{
rearing_time_yolo(model_day, sim_type)
}
\arguments{
\item{model_day}{Model day when rearing was initiated}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Yolo Bypass rearing time after adjusting indirectly for passage time and directly for temperature
}
