% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_one_rep.R
\name{run_one_rep}
\alias{run_one_rep}
\title{Run one replicate of the model}
\usage{
run_one_rep(
  water_year,
  chinook_run = c("Fall", "LateFall", "Spring", "Winter"),
  scenario = c("Exg", "Alt01", "Alt04b", "Alt04", "Alt05", "Alt06"),
  sim_type = c("deterministic", "stochastic")
)
}
\arguments{
\item{water_year}{Water year (1997-2011)}

\item{chinook_run}{Run timing classification: Fall, LateFall, Winter, Spring}

\item{sim_type}{Simulation type: deterministic or stochastic}
}
\description{
Run one replicate of the model; a replicate is a water_year/chinook_run/scenario/sim_type combination
}
