% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_adjustment.R
\name{temperature_adjustment}
\alias{temperature_adjustment}
\title{Adjust rearing time based on temperature}
\usage{
temperature_adjustment(model_day, duration, location)
}
\arguments{
\item{model_day}{Model day when rearing was initiated}

\item{duration}{Maximum numbr of rearing days}

\item{location}{Rearing location: Yolo or Delta}
}
\description{
Rearing time after potentially truncating based on temperature threshold
}
