% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toe_drain_temperature}
\alias{toe_drain_temperature}
\title{Toe Drain temperature}
\format{
A list of 3 vectors of 5477 elements:
\describe{
  \item{Date}{calendar date}
  \item{DOY}{day of year}
  \item{Value}{mean daily temperature (ºC)}
}
}
\usage{
toe_drain_temperature
}
\description{
A dataset containing the daily temperature in the Toe Drain at the screw trap
}
\keyword{datasets}
